server: Enable data_distribution API for app tenants.
Previously, this endpoint scanned over meta KVs and used their startKey to
determine the table they belong to. However, this approach does not work
for coalesced ranges.

The fix, inspired by the `SHOW CLUSTER RANGES` query, uses the
`crdb_internal.table_spans` table and joins it with the
`crdb_internal.ranges_no_leases` table.

There is still some work left that will be addressed in follow-up PRs:

1. While this fixes the behavior for secondary tenants, there is a known
   bug where the result will not contain deleted (not GC'ed) tables. This
   will be fixed in a follow-up PR by including deleted tables to the
   `table_spans` table and introducing an `is_deleted` column. It will be
   ensured that the behavior of `SHOW CLUSTER RANGES` doesn't change.
2. We might want to add a summation of ranges at various levels because the
   sum of replicas of all tables in a database can be more than the actual
   number of ranges that belong to that database due to coalescing.

Overall, this PR also fixes the broken behavior to a large extent of this
endpoint since range coalescing is enabled. But let's wait for "1" to be
finished before we declare this work as done.

Informs: #97942
Fixes: #109501
Epic: CRDB-12100
Release note: None